review: split oversized test files, verb-first stamp helper (#274)
Round 2 P2/P3 standards items. Three test files exceeded the 500-line
limit after this PR extended them; each is split along the seam this PR
touched, pure moves with assertions unchanged:
- export_sql_shared_test.go (503) -> mode-case type/fixtures to
  export_sql_mode_cases_test.go (418 + 92);
- postgres_persistent_repository_test.go (507) -> statement-builder and
  stamp tests to postgres_persistent_repository_statement_test.go
  (328 + 192);
- postgres_persistent_repository_repo_test.go (758, pre-existing
  violation extended here) -> the update/delete/batch-delete pgxmock
  suites to postgres_persistent_repository_write_mock_test.go
  (372 + 406).

tombstoneStamp is renamed computeTombstoneStamp (verb-first private
helper naming).
